+++ index.html
@@ -16,7 +16,9 @@
             <li><a id="list-link" href="#" class="nav-link-list">List</a></li>
             <li><a id="add-link" href="#" class="nav-link-list add">Add</a></li>
             <li>
-              <a id="contact-link" href="#" class="nav-link-list contact">Contact</a>
+              <a id="contact-link" href="#" class="nav-link-list contact"
+                >Contact</a
+              >
             </li>
           </ul>
         </nav>
